+++ lib/representable/deserializer.rb
@@ -5,6 +5,8 @@ module Representable
     end
 
     def deserialize(fragment)
+      # puts "deserialize #{@binding.name}" # TODO: introduce Representable::Debug.
+
       # next step: get rid of collect.
       fragment.enum_for(:each_with_index).collect do |item_fragment, i|
         @deserializer = ObjectDeserializer.new(@binding)